+++ pycktestproject/pycktestproject/apps/__init__.py
@@ -1,0 +1,13 @@
+# This list contains all the enabled apps for the project, when you add a new app, add it to this list
+# to make it accessible by the project.
+
+import importlib
+
+PROJECT_NAME = 'pycktestproject'
+project_package = importlib.import_module("pycktestproject")
+
+#from . import app1, app2
+#enabled_apps = [app1, app2]
+
+enabled_apps = []
+