+++ pycktestproject/pycktestproject/models/models.py
@@ -1,0 +1,10 @@
+from sqlalchemy import (
+    Column,
+    Integer,
+    Unicode,
+    )
+
+from . import db, Base
+
+
+# Create your models here.