+++ docs/changes.rst
@@ -10,6 +10,7 @@ What's new in 0.9.5
 
 * Bugfix: PyCK auth controllers used request.current_route_url() with query strings that caused duplicate actions, like delete a record and then redirecting to the same url (with action delete) and trying to delete it again.
 * Bugfix: In admin section if an integer primary key is None (empty), it tries to insert or update that as a string 'None' causing error.
+* Allow redirecting from CRUD Controller add, update and delete actions.
 
 What's new in 0.9.4
 -------------------